For 32 bit systems, cast the long long used by strtoll() to unsigned long and then a void * for calling the hexdump() operator with bluhm

+++ sys/stand/boot/cmd.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: cmd.c,v 1.65 2019/08/03 15:22:19 deraadt Exp $	*/
+/*	$OpenBSD: cmd.c,v 1.66 2019/11/28 00:17:13 bluhm Exp $	*/
 
 /*
  * Copyright (c) 1997-1999 Michael Shalayeff
@@ -368,7 +368,7 @@ Xhexdump(void)
 			return 0;
 		}
 	}
-	hexdump((void *)val[0], val[1]);
+	hexdump((void *)(unsigned long)val[0], val[1]);
 	return 0;
 }
